Refactor picom config to use template
My old picom config was an amalgamation of lines I found on r/unixporn
and dotfile repos. It was ugly and hard to maintain, so I made the
effort to port what I had to the (relatively) new picom configuration
template file.

Most of the settings were surprisngly the same as default, with a few
tweaks here and there.